requirements.txt
fastapi
uvicorn
pillow
numpy
scikit-learn
xgboost